soc/intel/apollolake/acpi: Remove the PMC IPC asl
The PMC (0d.1) can't be enumerated or probed, as the P2SB (0d.0)
is hidden by coreboot.

That means the IPC1 driver won't ever be able to communicate with
the PMC. Remove the non-functional driver to stop it creating errors
in both Linux and Windows.

The following error can be observed on Linux:
   intel_punit_ipc intel_punit_ipc.2.auto: Invalid IRQ, using polling mode
   intel_punit_ipc intel_punit_ipc.2.auto: invalid resource


Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Change-Id: Idf87b0d983deda6ed1ac0306a8e60a9bbb5276fd