drivers/efi/efivars.c: avoid overwriting rdev
Implementation of efi_fv_*() functions modifies passed in
`region_device` as an undocumented side-effect.  This behaviour is
highly unexpected and can break callers in unobvious ways.

Keep the implementation as is, but make a copy of the `region_device` in
public functions before doing any processing.

Change API to accept constant `struct region_device` so that future
change can't reintroduce this bug by using `rdev` instead of
`&store_dev`.

Change-Id: Ie72df151088b80f8e85e9eba149a3f99ce5446ee
Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/94280
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-by: Patrick Rudolph <patrick.rudolph@9elements.com>