drivers/payload_mm_interface: Implement payload MM concept
Implement support for coreboot as the bootloader in the payload MM
concept, as described by the documentation.

The payload MM feature requires further changes to be made to the tree.
So, introduce the `HAVE_PAYLOAD_MM_INTERFACE` Kconfig symbol which will
be selected by platforms that have implemented said changes.

Change-Id: Ic1e0a88be1f74a2182256d2a03e7266847e63300
Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>
Signed-off-by: Angel Pons <th3fanbus@gmail.com>