i2c/drivers/generic: Add support for including a rotation matrix
The Rotation Matrix' allows specifying a 3x3 matrix representing
the orientation of devices, such as accelerometers. Each value in
the matrix can be one of -1, 0, or 1, indicating the transformation
applied to the device's axes.

It is expected by linux and required for the OS to correctly interpret
the data from the device.

Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Change-Id: Id4a940d999a0e300a6fe21269f18bab6e3c0523c