soc/intel/alderlake: Add helper functions for Power Management
Clock Power Management as ASPM is configured inconsistently on
Alder Lake, depsite the UPDs remaining the same since Skylake.

For most SOCs, the Clock Power Management UPD isn't touched, so even
if coreboot is configured to disable it, FSP can re-enable it.

ASPM is configured from most SOCs, but not for Alder Lake.

Add helper functions to configure this correctly.

Change-Id: I9b0d15d8783548b16a3726a5d191fc4042eb56ae
Signed-off-by: Sean Rhodes <sean@starlabs.systems>