soc/intel/cnvi: Add CNVW OpRegion
The CNVi driver is relatively basic in coreboot and most
noticeably, recent Linux kernels flag that lack of a _PRR
method, which is used to reset WiFi and Bluetooth.

This patch series adds methods recommended by Intel in
document #559910.

This patch defines an OpRegion for CNVi, for both
integrated and dedicated solutions.

Change-Id: Idd2ff93fb65c40f656804d96966e1881202ccb56
Signed-off-by: Sean Rhodes <sean@starlabs.systems>