ec/starlabs/merlin: Add Intel Virtual Button driver
Add Intel Virtual Button driver which is used to report
to the OS whether a tablet is docked or undocked.

This is currently only used on `mb/starlite_adl` so the
GPIO is hardcoded to GPP_F15 for now.

Tested on `starlite_adl` with Ubuntu, by checking the
virtual keyboard is show when the tablet is undocked
and hidden when docked.

Change-Id: I574a1b2d3907b2341a0dfdc412151d574ba4848e
Signed-off-by: Sean Rhodes <sean@starlabs.systems>