{arch,cpu}/x86, console: Add support for debug register exceptions
coreboot's breakpoint-based exceptions can be used during SMM runtime,
but some considerations are needed. Architecturally, only DR6 and DR7
are saved/restored by the SMM save state, which means that DR0-DR3 must
be saved and restored by the SMI handler. Additionally, the runtime is in
an interesting position where variables remain initialised, but
the CPU state resets. Therefore, we have to formally remove
the NULL and stack canary breakpoints at the end of SMI handling too.

Since the interrupt handler could be the GDB stub, and developers might
also want to debug SMM code, also hook up the common stub to the UART
transport.

Also take the opportunity to do some cleanup: `exception_init` is called
directly, so make sure the prototype is in the header (not the definition).

Change-Id: I3445cb7cc4079ce9ad840b6a1e37a06dcb9abb74
Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>