Documentation/drivers: Add payload MM documentation
Payload MM is a new feature that enables payloads to execute in SMM.
An example use-case for this, and the primary one being enabled, is
UEFI Secure Boot.

As I describe in the blog post we're releasing together with this
patchset, it is critically important that writes to those variables
be cryptographically verified against their PKI. This means that
enabling authentication in UefiPayload and SMMSTORE in coreboot
is insufficient, since the SMMSTORE driver could be called out of band.

Therefore, let coreboot own SMM, and let the payload own its features.

Change-Id: Ifa6160d6a09e646e858932a0ce64e2c42935604e
Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>
Signed-off-by: Angel Pons <th3fanbus@gmail.com>