drivers/payload_mm_interface: Support variable services use-case
This passes the needed SPI controller information in the hand-off tables
for Intel SoCs. QEMU flash devices are supported solely using the FMAP,
and AMD SoCs are unsupported at this time.

This patch also makes the payload MM interface driver an equally capable
solution with the same user experience as SMMSTORE.

Change-Id: I39b0837fafc81afe80f19f618783be8becf7353c
Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>
Signed-off-by: Angel Pons <th3fanbus@gmail.com>